ANOVA method for a fitted serp object

Provides a likelihood ratio test for comparing two or more serp objects. This does not currently support model(s) with penalized slope.

# S3 method for serp
anova(object, ..., test = c("Chisq", "none"))

Arguments

object

An object of class serp.

...

additional arguments.

test

type of test to be conducted.

Value

model

the respective model aliases.

slope

type of slope fitted, which may be any of, unparallel, parallel, or partial slope.

no.par

the no of parameters in the model.

AIC

the akaike information criterion.

logLik

the realized log-likelihood.

Test

the different pair(s) of test(s) conducted.

LR.stat

the computed Likelihood ratio statistic.

df

the degree of freedom.

Pr(chi)

the p-value of test statitic.

Details

An ANOVA table with the following components on display:

See also

Examples

library(serp)
m1 <- serp(rating ~ temp + contact, slope = "parallel", link = "logit",
           data = wine)
m2 <- update(m1, ~ contact)
anova(m1, m2)
#> Likelihood ratio tests of ordinal models:
#> 
#>   Model    slope no.par    AIC  logLik   Test LR.stat df   Pr(Chi)
#> 1    m2 parallel      5 209.91 -99.956                            
#> 2    m1 parallel      6 184.98 -86.492 1 vs 2  26.928  1 2.112e-07