Bitterness of wine dataset

The wine dataset adopted from Randall(1989), represents the outcome of a factorial experiment on factors determining the bitterness of wine. Two treatment factors (temperature and contact) with two levels each are provided, with the rating of wine taken on a continuous scale in the interval from 0 (none) to 100 (intense). These were subsequently grouped into five ordered categories ranging from 1 = 'least bitter' to 5 = 'most bitter'. Altogether, nine different judges assessed wine from two bottles and out of the four treatment conditions, making a total of 72 observations.

wine

Format

A data frame with 72 rows and 6 variables:

response

scorings of wine bitterness on a 0---100 continuous scale.

rating

ordered factor with 5 levels; a grouped version of response.

contact

factor with two levels ("no" and "yes").

temp

temperature: factor with two levels.

judge

factor with nine levels.

bottle

factor with eight levels.

Source

Taken from Randall (1989).

References

Randall, J (1989). The analysis of sensory data by generalized linear model. Biometrical journal 7, pp. 781--793.

Examples

if (FALSE) {
str(wine)
head(wine)
}