Performance metrics for categorical models

Calculates performance metrics of fitted categorical models, including binary and multi-categorical models.

errorMetrics(
             actual,
             predicted,
             model= c("multiclass", "binary"),
             type= c("brier", "logloss", "misclass"),
             eps=.Machine$double.eps)

Arguments

actual
vector of actual values observed

predicted
predicted probability matrix of a categorical model or a vector of fitted values for binary models.

model
specifies whether multi-categorical or binary model

type
specifies type of error metrics

eps
a near-zero value introduced only if the fitted probabilities go beyond a specified threshold. It helps to minimize the chances of running into numerical problems.

Value

A numeric value of computed performance metric determining how good a categorical model is compare to competing models.

brier

the brier score of fitted model.

logloss

the logloss of fitted model.

misclass

the misclassification error of fitted model.

See also

Examples

f1 <- serp(rating ~ temp + contact, tuneMethod = "user",
slope = "penalize", lambda = 0.3, reverse = TRUE, link = "logit",
data = wine)
errorMetrics(f1, type = "brier")

#> [1] 0.6613498

errorMetrics(f1, type = "logloss")

#> [1] 1.18257

errorMetrics(f1, type = "misclass")

#> [1] 0.7083333

## For non-serp object of class, 'actual'  and 'predicted' values
## must be provided
set.seed(1)
y <- as.factor(rbinom(50,1,0.5))
xx <- runif(50)
f2 <- glm(y ~ xx, family = binomial(link="logit"))
p2 <- f2$fitted.values

errorMetrics(actual=y, predicted=p2, model= "binary", type = "brier")

#> [1] 0.244247

errorMetrics(actual=y, predicted=p2, model= "binary", type = "logloss")

#> [1] 0.6818895

errorMetrics(actual=y, predicted=p2, model= "binary", type = "misclass")

#> [1] 0.54