Summary method for a fitted serp object.

This function summarizes the result of a fitted serp object in a dataframe.

# S3 method for serp
summary(object, ...)

Arguments

object
An object of class serp.

...
Not used. Additional summary arguments.

Value

an object of class summary.serp. A list (depending on the type of slope used) of all model components defined in the serp, function with additional components listed below.

coefficients

the matrix of coefficients, standard errors, z-values and p-values.

null.deviance

the deviance for the intercept only model.

null.logLik

the log-likelihood for the intercept only model.

penalty

list of penalization information obtained with slope set to "penalize".

expcoefs

the exponentiated coefficients.

See also

Examples

library(serp)
m <- serp(rating ~ temp + contact, slope = "penalize",
           reverse = TRUE, link = "logit", tuneMethod = "user",
           lambda = 0, data = wine)
summary(m)

#> 
#> call:
#> serp(formula = rating ~ temp + contact, link = "logit", slope = "penalize", 
#>     tuneMethod = "user", reverse = TRUE, data = wine, lambda = 0)
#> 
#> Coefficients:
#>                Estimate Std Error z value Pr(>|z|)    
#> (Intercept):1    1.2258    0.5570   2.201 0.027757 *  
#> (Intercept):2   -1.0330    0.4807  -2.149 0.031638 *  
#> (Intercept):3   -3.9464    0.9015  -4.377  1.2e-05 ***
#> (Intercept):4  -19.1843 1079.7764  -0.018 0.985825    
#> tempwarm:1      19.2427 3597.5553   0.005 0.995732    
#> tempwarm:2       2.1108    0.6007   3.514 0.000441 ***
#> tempwarm:3       2.9404    0.8283   3.550 0.000385 ***
#> tempwarm:4      17.0636 1079.7763   0.016 0.987392    
#> contactyes:1     1.6594    1.1786   1.408 0.159167    
#> contactyes:2     1.3429    0.5830   2.303 0.021258 *  
#> contactyes:3     1.6928    0.6602   2.564 0.010347 *  
#> contactyes:4     1.1622    0.9054   1.284 0.199293    
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Number of iterations: 18 
#> 
#> Loglik: -84.61093 on 276 degrees of freedom 
#> 
#> AIC: 193.2219
#> 
#> Exponentiated coefficients:
#>   tempwarm:1   tempwarm:2   tempwarm:3   tempwarm:4 contactyes:1 contactyes:2 
#> 2.275154e+08 8.255243e+00 1.892362e+01 2.574225e+07 5.256007e+00 3.830073e+00 
#> contactyes:3 contactyes:4 
#> 5.434719e+00 3.196963e+00 
#> 
#> Regularization Info:
#> penalty:     SERP
#> tuneMethod:  user
#> value:        
#> lambda:      0