Summary method for a serp object.

Summarizes the results of the fitted model in a dataframe.

# S3 method for serp
summary(object, ...)

Arguments

object
An object of class serp.

...
Not used. Additional summary arguments.

Value

an object of class "summary.serp", a list (depending on the type of slope used) with the components itemized below. Note that the 'components from object' are already defined in the main function.

call

the component from object.

link

the component from object.

edf

the component from object.

ylev

the component from object.

nobs

the component from object.

gradient

the component from object.

Hessian

the component from object.

fitted.values

the component from object.

slope

the component from object.

Terms

the component from object.

control

the component from object.

reverse

the component from object.

converged

the component from object.

iter

the component from object.

message

the component from object.

misc

the component from object.

model

the component from object.

coefficients

the matrix of coefficients, standard errors, z-values and p-values.

logLik

the component from object.

deviance

the component from object.

aic

the component from object.

bic

the component from object.

contrasts

the component from object.

penalty

list of penalization information obtained with slope set to "penalize".

expcoefs

the exponentiated coefficients.

cvMetric

the component from object.

globalEff

the component from object.

lambda

the component from object.

lambdaGrid

the component from object

na.action

the component from object.

nrFold

the component from object.

testError

the component from object.

tuneMethod

the component from object.

value

the component from object.

See also

Examples

# See serp() documentation for examples.